Cambios Usuario y ProfileController

+++ app/Http/Controllers/Auth/LoginController.php
@@ -10,39 +10,46 @@ use App\Models\User;
 class LoginController extends Controller
 {
     public function login(Request $request)
-{
-    $request->validate([
-        'email' => 'required|email',
-        'password' => 'required',
-    ]);
+    {
+        $request->validate([
+            'email' => 'required|email',
+            'password' => 'required',
+        ]);
 
-    if (!Auth::attempt($request->only('email', 'password'))) {
-        return response()->json(['message' => 'Credenciales incorrectas'], 401);
+        if (!Auth::attempt($request->only('email', 'password'))) {
+            return response()->json(['message' => 'Credenciales incorrectas'], 401);
+        }
+
+        $user = User::where('email', $request->email)->first();
+
+        // 🚫 Validación del estado del usuario
+        if ($user->estado == 0 || $user->is_deleted) {
+            return response()->json(['message' => 'Este usuario está inactivo o eliminado.'], 403);
+        }
+
+        // Eliminar todos los tokens anteriores del usuario
+        $user->tokens->each(function ($token) {
+            $token->delete();
+        });
+
+        // Crear un nuevo token
+        $token = $user->createToken('api-token')->plainTextToken;
+
+        // Cargar relación 'role' (esto es lo nuevo)
+        $user->load('role');
+
+        return response()->json([
+            'token' => $token,
+            'user' => [
+                'id'    => $user->id,
+                'name'  => $user->name,
+                'email' => $user->email,
+                'role'  => $user->role->id, // Aquí aseguramos enviar el nombre del rol
+                'roleName'  => $user->role->nombre, // Aquí aseguramos enviar el nombre del rol
+                'image' => $user->image // Asegúrate de que la imagen esté disponible
+            ],
+        ]);
     }
-
-    $user = User::where('email', $request->email)->first();
-
-    // Eliminar todos los tokens anteriores del usuario
-    $user->tokens->each(function ($token) {
-        $token->delete();
-    });
-
-    // Crear un nuevo token
-    $token = $user->createToken('api-token')->plainTextToken;
-
-    // Cargar relación 'role' (esto es lo nuevo)
-    $user->load('role');
-
-    return response()->json([
-        'token' => $token,
-        'user' => [
-            'id'    => $user->id,
-            'name'  => $user->name,
-            'email' => $user->email,
-            'role'  => $user->role->id, // Aquí aseguramos enviar el nombre del rol
-        ],
-    ]);
-}
 
     public function logout(Request $request)
     {
@@ -54,4 +61,3 @@ class LoginController extends Controller
         return response()->json(['message' => 'Sesión cerrada']);
     }
 }
-
